Add option to sort data by x-axis for vtkPVXYChartView
The sorted order of the data arrays determines the connectivity of the
line graphs.  And many line graphs want the line to be connected along
the points in order of their x-coordinate.  This option allows this type
of graph even when the original data is not sorted in that order.