When wrapping prefer to exclude files before checking if they exit
When you are building up the list of headers for client server wrapping,
we would previously check for file existence before exclusion. This
caused problem when a file couldn't be found because the search paths
didn't include private sub folders. Now the problem is avoided as those
files are on the excluded list.