Add a new AppendReduce Filter
ParaView miss a usable and exposed ReductionFilter
This commit modify the reduction filter
to make it more usable while keeping it's default behavior
for rendering usage. It also expose a AppendReduce
filter which by default Reduce all dataset on all nodes
using an AppendFilter.
This also add a test for the new filter.