Enable OSPRay progressive refinement
Run paraview with --enable-streaming (client and server) and then
change "max frames" on the OSPRay view controls to something
larger than 1 to use it. When you do, still renders run in a loop,
until interrupted by camera motion, and OSPRay accumulates in
new renderings with slightly difference sub pixel origins. The end
result is the same as cranking up the samples per pixel and AO
samples, but it is more reponsive to the user.