Split Truchas reader test into two versions
One version is for when VisItBridge is enabled, one version is for
when VisItBridge is not enabled. VisItBridge adds several readers that
say they can read .h5 files like the Truchas reader, hence the test
needs to pick the Truchas reader from the reader selection dialog.
Without VisItBridge enabled, the Truchas reader will be used
automatically.