Remove NO_OUTPUT option from halo finder tests
The NO_OUTPUT option prevents tests from receiving the -T option on
the command line that specifies the output directory. When one of
these tests failed, vtkTesting would try to save the test-generated
images and diff images to
<build>/ParaViewCore/VTKExtensions/Testing/Temporary/. This directory
does not exist, hence error messages that the test images could not be
written to the temporary location.

Removing the NO_OUTPUT option makes the testing infrastructure use the
correct temporary output directory.