Minimize paints after resize with Qt 5.
With `QVTKOpenGLWidget`, we can avoid an extra render request on the
render window after the widget has been resized by using
`QOpenGLWidget::resized` signal to update the size properties on the
proxy. Thus, we're ensuring that the properties are updated before the
paint request in resize event gets triggered, but after the widget's
size parameters have been updated.