Don't use single array property for repeatable property
This was causing exceptions in the python tracing code when there
happened to be 5 elements in one of these repeatable properties.  These
were selection properties where the user could select multiple arrays
and the values were a list of array names.  Trying to connect these
names to int in ArraySelectionProperty was causing exceptions and making
python trace fail.

See #16906.