Fix scalar bar placement in Python state.
The `Orientation` property needs to be set before the `Position*`
properties on the scalar bar for it to work correctly. While this
requirement is absurd, it's the legacy scalar bar! Hence just fixing the
Python state to save out the order correctly. This is done by simply
reordering the properties in XML since the Python state preserves XML
order for properties.

Addresses #17197.