Fix tracing/saving state of representation property.
Addresses #17196.

The issue is a tricky one. Here's the synopsis. When tracing or saving
state in Python, we only save state for properties that are different
from the default. This generally is a good strategy for minimizing trace
verbosity. Now, for a property with `vtkSMRepresentationTypeDomain`, the
default is changed by the domain (in our case to `Outline` from
`Surface`). For the bug, we changed it back to `Surface` -- now same as
default. Hence, when saving the state file, the code skipped the
property as it's value was unchanged from the default -- causing the
bug.

As a fix, we force the tracing mechanism (also used in when saving state
as Python) to save the "Representation" property, even if it hasn't
changed from the default value.

A better fix maybe to add support to vtkSMProperty to know that it was
changed since initialization -- but that needs some pondering.