Fix issues with collaboration tests.
vtkPVDataDeliveryManager only needs to use the information about known
representations, and not necessarily ports since that info is no longer
available until a representation executes, which need not be similar on
all clients e.g. is the master changes active representation type and
a slave joins after this change, it will now know of the old
representation (e.g. pvcs-collab.VariableSelection1 test).

Fixed that.