treamLinesRepresentation: Various fixes and improvements.
1/ Address a bug when using thick lines (width>1) with GL 3.2 core profile.
With this profile (used on some systems), glWidthLine() does not
have effect and generate an error: one must process line thickness
manually. A striped version of the geomtry shader of the
OpenGLPolyDataMapper has been reproduced here.

2/ Change the blending equation to support any background color.

3/ Support for actor/user transform.

4/ Add a new "Animate" option to turn the animation ON/OFF.

5/ Tests are also introduced thanks to a new optiow "NumberOfAnimationSteps"
which perform a single rendering pass with the given number of steps
and then stop updating.