Refactor "Save Screenshot".
Refactoring the infrastructure to save screenshots in ParaView. These
changes move logic to capture a screenshot of abitrary resolution to
ServerManager. Using a proxy for specifying the options for saving a
screenshot, all aspects of that operation -- including changing color
palette, stereo mode, etc. -- are now accessible in ServerManager and
hence in Python.

Also options relevant to saving screenshot e.g. view border,
transparency for images, are now shown in the save screenshot dialog
itself, without having to change the same in the "Settings".

Adding UI to control font-scaling when saving screenshots allows uses to
scale fonts (for high-PPI outputs) or maintain size (for
high-resolution, same PPI outputs).

Addresses #17223, #16107, #14955, #14958.