Cleanup server-side proxies on disconnect.
vtkSMSession was stopping clearing of server-side proxies on disconnect.
This was only needed in collaboration mode, doing it all cases  --
while okay for servers that are going to quit anyways -- has
consequences for actions like disconnect-and-save animations. Fixed to
only disable sending of cleanup messages to server when the client is in
collaboration mode.