Refactor "Save Animation".
Refactoring the infrastructure to save animations as images/movies in
ParaView. Moving the logic to ServerManager, and using the existing (and
newly refactored) infrastructure to save screenshots for views and
layouts. This makes screenshot and animation options consistent.

Also refactored support for disconnect-and-save animation. To keep
things simple, we now simply recreate client state on the server after
disconnect (rather than trying to reuse pipeline proxies). The
complication needed for that was not worth the minor savings in
read/compute time for the single timestep.

Addresses #14912, #14958, #14955.