Avoid aggressive `_remove_functions`.
When a connection was disconnected, _remove_functions would remove all
functions using proxy names for all proxy modules known to
servermanager, even those that were never imported. That's bad since it
removes functions like `SaveScreenshot` which conflict with name of a
proxy in the options module that is not imported in simple namespace.
Fixed `_remove_functions` to only remove those function names for
proxies it actually imported.