Implemented 'smart lookup' for state data files.
Implemented 'smart lookup for state data files. This does two things.
First if the data file is not found in the user specified directory it starts
adding the original directory path starting the farthest down the tree
between the user specified directory and the filename until the file is
found or the original path has been exhausted. Second it keeps track of
how many times the parent directories of the new file location match for
file from a single property element. Once the number of matches reaches a threshold
it assumes the parent directory will apply to the remaining files for that
property element. Also fixed a bug where file path was updated for only
the first file in a series of files in the same property element.