Cleanup capture code in View proxies.
Cleaning up image capture code in vtkSMViewProxy and subclasses. Now, a
we provide a capable default implementation in vtkSMViewProxy itself.
Thus, removes duplicate(ish) code for several subclasses.

Next, we use an internal subclass of vtkWindowToImageFilter to call
`vtkSMViewProxy::StillRender` when the filter wants to request a render
for saving images. That overcomes the dependence on the interactor, thus
overcoming deadlocks when the interactor was missing (issue #17205).