Fixed issues from code review and other feedback.
These include using xpath for more efficient parsing of the state
file xml. Moving where state file parent directory is set as the
default for the directory selection part of the dialog so the
directory path actually appears in the selection box. Also switched
the filepaths data structure from a map back to a vector.