Add a "Link Selected Elements" to Selection Link
This add a Link Selected Elements mode to Selection link
which enable to convert any selection to a ID based selection
before appplying selection on linked datasets.
This also had a dialog to select the option
as well as a checkbox to check/uncheck in the
link editor.
This add tests for this new feature as well.
This also correct a small ordering bug in the SelectionLinkReaction