Explicit file replacement for load state working.
The explicit file replacement for the laod state dialog is working.
Required creating new proxies for each source with data files to add them
as sub-proxies to the LoadStateOptions proxy.
The explicit file listing is visible only when the correct option
is selected. Tested the file
explicit replacement on several test cases and it works. Also
refactored the code that parses location enviroment variable so
it will work with any environment variable.