fix first frame with OSPRay in parallel (and thus Catalyst)
Before, the following script would hang when run with mpiexec -np 2 bin/pvbatch,

from paraview.simple import *
v = CreateView('RenderView')
v.EnableOSPRay = 1

unless a call to Render() was put in before the switch to OSPRay.

Removing the this->Render() call in vtkPVRenderView fixes that, and it does not
seem to be necessary anyway.