Add ability to show "Plot Over Line" in active view.
Adding mechanism for filters to request showing their result in the
active view if possible, besides the suggested preferred-view. This
ensures that the line from "Plot Over Line" is shown in the active
Render View too thus giving the user an indication of where the probe is
happening.

Updated ExportLinePlotToCSV to test this.

Fixes #17146.