Fix Toggle Connectivity feature in SpreadSheet view.
Fixes #15749.

So far, the feature only had an effect on the first representation.
This change fixes that and applies it any selected source output port.
Add a GenerateCellConnectivity to the spreadsheet view to make the
feature global to the view and be able to restore it state from file.
Fix the issue where the spreadsheet UI was not reflecting the visible
output when loading back a state file.

Add leading zeros (0) to "Point Index" arrays to avoid annoying sorting
effect when there are more than 10 points per cell.