Replace Position2 property with ScalarBarLength
A separate ScalarBarLength property is much more intuitive than the
Position2 property that required the right element to be set depending
on the orientation. Also, Position2 is a misnomer - it is a relative
position, but really it specifies the size of the scalar bar.

Unfortunately, the Position2 attribute is deep down in the class
hierarchy, so we are stuck with it when interacting with the scalar
bar widget as it stands. This commit adds necessary translation code
that synchronizes the ScalarBarLength property with the Position2
property.