Add WindowLocation parameter to state file XML
By default, the WindowLocation for scalar bars is the lower right corner.
This makes it so that any state file saved from ParaView 5.3 earlier
sets the scalar bar WindowLocation to "AnyLocation" so that the Position
property can be respected. Otherwise, all the scalar bars from older
state files pile up in the lower right corner.