Possibility to turn off automatic display of filter output.
After 'Apply' ParaView displays all outputs of a source/filter in the selected View or in a new View.
The user may want to disable this automatic display for a specific output port of the source/filter.
This modification allows the user to use the SourceProxy hint 'View' with attribute 'type="None"',
which will prevent the automatic display.