Fix flash when selection blocks in render view.
The flash was happening because Qt requested a paint during end of
selection stages on QVTKOpenGLWidget as ParaView was painting the
progress bar (of course!). Since ParaView thinks its busy (since
progress state is in the process of being toggled off), it decided to
skip the render request. Since QVTKOpenGLWidget had no mechanism to not
copy VTK's FBO over to the screen, it went ahead and did that, causing
the flicker. Fixed that.

Fixes paraview/paraview#17256.