When appending attributes mangle array names if the name is repeated
For repeated array names, previously the last array would be the
one that would be output in the vtkMergeArrays filter
(Append Attributes). Change to mangle the name if the array name
is repeated. The mangling appends the input array name with
"_input_<index>" to help guide the user as to which input
the array came from. A global consistency check is also now
performed for parallel use of the filter. This should now also
work with all vtkDataObjects including multiblock data sets.

Fixes #17323