BUG: Cannot do fresh CMake configuration on macOS
The variable PARAVIEW_QT_VERSION was wrongly assumed to exist when it
was checked to determine whether ParaView can enable high resolution
mode on Mac builds. In fresh builds of ParaView, this led to a syntax
error in an 'if' statement and prevented configuration.

Moving the high resolution check to after the setting of
PARAVIEW_QT_VERSION solves the problem.