Avoid remote rendering when making selections if not needed.
When making hardware selection in client-server mode, with local
rendering, we don't need any rendering stages to happens on the server
side. Rendering for selection buffers can happen on client alone. We
were requesting capturing of selection buffers (and as a consequence,
rendering) on all ranks. This was causing BUG #17416, as ParaView was
triggering redistribution code erroneously. Fixed that.

vtkSMRenderViewProxy ensures that the "Select*" calls are now only made
on rendering processes, similar to StillRender and InteractiveRender
calls. This should speedup selection in client-server mode when local
rendering.

Fixes 17416