Avoid changing model on treeview.
Seems like QTreeView is not robust when model on it keeps changing.
After the model is changed, things like updates to the view when
checkstate of a node was toggled or color being changed weren't getting
reflected promptly on the view.

Instead we set/unset the sourceModel on the proxyModel. That works more
reliably.