Label, tick, and title placement improvements
Based on community feedback, tweak the tick marks to make them not
appear to be negative signs in front of labels. Also, base the tick
length and spacing between tick marks and labels on the label font
size, and base the spacing between the title and labels on the title
font size. These tweaks improve the aesthetics of the color legend.