Ensure that all representations are up to date for InteractiveRenders
In the past, we avoided updating the view during InteractiveRenders to
improve performance. These days, the view keeps track of whether
anything needs to be updated and Update() returns immediately if no
updating is necessary, hence we can call it within InteractiveRender()
and it causes no degredation in the interactivity.