Modify transfer functions range reset logic
The transfer function range reset mode is now applied per color map in
vtkSMTransferFunctionManager::ResetAllTransferFunctionRangesUsingCurrentData.
The logic for determining when and how to reset has been consolidated
in this member function.

The meaning of the boolean second parameter of this function has
changed: It now indicates whether the function is being called in the
context of animation. Because the reset modes change the resetting
behavior based on whether animation is occuring or not, we must supply
this function with the animation state.