Introduce animation shortcuts
This commits adds a new AnimationShortcut Decorator and Widget for some properties.
The widgets (buttons with a Play, Ramp or Delete icon) are visible if a global setting "ShowAnimationShorcuts" is turned ON.
Clicking on such a button adds/edit/delete a new animation track for the associated property and opens a dialog to specify the keyframes parameters as long as the number of frames for the animation.