Improve image size/scaling code.
Improved code in vtkSMSaveScreenshotProxy that determines the scale
factors to use to achieve a target resolution when saving images. The
new logic allows for independent magnification factors which is crucial
to provide accurate target resolution.

This caused API on vtkSMViewProxy (and subclasses) to change, since now
`CaptureWindow` call takes in a 2-component magnification factor instead
of a single value.

Addresses #17567.