Set AutomaticRescaleRangeMode to global value when resetting to data range
When a transfer function's range is set to the data range, reset the
AutomaticRescaleRangeMode to the global setting value. This mimics the
previous behavior of this function which disabled locking of the
scalar range, but updates it to account for the transfer
function-specific automatic rescaling mode.

In addition, skip resetting the range if the AutomaticRescaleRangeMode
is set to vtkPVGeneralSettings::NEVER instead of checking the old
LockScalarRange property.