Fix code removing proxy from menu
We used to need to remove proxy when it was unavailable server side
With the use of ParaViewFilters.xml and current design
menu are populated with an action only if a prototype can be built.
remove the proxy from the list is unnecessary and can lead to missing
a proxy when connecting to servers with different configuration.