Force Axes Grid to be opaque when rendering in symmetric mode
The Axes Grid labels are translucent, hence by default the label
fragments do not get written to the depth buffer and therefore IceT
skips them when compositing in symmetric mode. This change forces the
entire Axes Grid actor to be opaque so that the labels are composited.

Also, added test for batch mode Axes Grid rendering.

Fixes #17472.