Fix creation of property group decorators.
Property group can have decorators as any common property widget. However,
creation of the group decorators in pqProxyWidget is coded wrongly:
Only the very pqPropertyWidgetInterface used to create the property group
is queried for creation of the group decorator. But decorators are (mostly,
if not exclusively) coded in separate classes. For this reason, another
nested loop over property widget interfaces is needed to create all requested
decorators for every property group.