New setting to force single column menus for sources & filters
The problem concerns potential multi-columns menus larger than the
screen width that can be created by Qt - especially for the
Alphabetical filters menu.

A new advanced option ForceSingleColumnMenus is added in the General
Settings to force the sources/filters menus to be presented as a single
scrollable column to ensure that all the filters are selectable on low
resolution screens. Option is disabled by default.

Side effect: also fixes the ugly presentation of General Options.