Export color legend screenshots when exporting X3D files
Color legends were not exported in an X3D export. This patch adds
support for exporting PNG files containing images of the color legends
in the scene. The exported images have a transparent background, and
other data set representations and color legends will not appear in
the image.

The exported PNG images are not referenced in the exported X3D
file. They are instead intended to be loaded and displayed by an X3D
application.