Add class to compute transparent images
vtkImageTransparencyFilter takes two inputs, one of which is a scene
rendered with a white background and the other is the same scene
rendered with a black background.  It looks at co-located pixels from
each input image. If they are not the same color, the pixel must have
been from a transparent object. At such pixel locations in the inputs,
the transparency value in the output image is derived from the
difference of the HSV colors from the white- and black- background
images.

The first input must be the white-background image and the second
input must be the black-background image.