Improve the preview mode
Create a property named PreviewMode in vtkSMViewLayout to enter/exit
preview mode. Now we are able to enter (if we are not already in) the
preview mode automatically when saving screenshot or animation.
The target preview resolution will be used by default in saving
screenshot or animation dialog even if it is too high to display on the
screen. In addition, with the new pqSplitter, the splitter can be
painted even when the handle width is 0, so the previous workaround
method in pqMultiViewWidget is removed.