Add option --disable-further-connections
In multi-clients mode, the master client can enable or disable the further
connections from the collaboration panel. Default is "enable".
With option --disable-further-connections, initial state is "disable".

The master client can change the connect-id of the server.

Connected clients are not disconnected when mode or connect-id is changed
in the server.