Add an option to the Clip filter to clip exactly for boxes
The exact box clip can only clip to the interior of the box and has to
do that with 6 separate plane clips so it is very expensive. The GUI
could use a bit of cleanup to reflect that the exact option for the
box clip only works when crinkle clip is off and inside out is on.

Addresses #17438