Lights: add 'sync to camera' button
Add a custom button as a subclass of pqCommandButtonPropertyWidget,
which sets the light's position and focal point based on it's
type and the camera's position.

Add a 'Modified' method to vtkSMPropertyHelper, because the renderer
can change the vtkLight's position without the proxy knowing, so
the position must always appear modified.