make material library consistent among processes
No longer using an actual singleton. However we still
effectively have one as we make one instance and assign it
to all views and representations.

We always read the material library on the server and
then serialize the content back to the client so everyone can
agree. The library search path is modelled after plugins.

We update the material domain that objects can
choose to color by via domains that watch the material library.

The UI to load additional materials is now on the menu not the
view.