LightProxy: when light type changes, update position/focal point
When changing from a headlight to a scene light, the renderer has
set the headlight's position/focal to match the camera. Get the
position/focal and set them in the proxy for the scene light.

For a camera light, reset to a default position/focal, so scene
coords aren't used for the relative position/focal of a camera light.