Fix multiply defined section name warnings
The doxygen "section" command has the format:

\section <section-name> (section title)

Some section declarations supplied only the title portion, and not the
section name. Hence, the first word of each title was parsed as the
section name. This resulted in multiply defined section names or the
first word of the title being swallowed in the generated
documentation. Fixed both errors by giving unique section names to
each declaration.