Better automatic category generation
This change the behavior of Annotations and color generation
for category.

This ensure the categorical table will never go down in size
when automatically setting anotations or using a color preset.
This also uses a modulo on the color preset to give color to automatic
annotations and when using color preset.

This adds the possibility to ovverride MaxDiscreteValues
when evaluating ProminentValue for category generation

This adds a default color preset (KAAMS)