Fix resetting of default color maps
Resetting a color map did not work well. If a site setting was
present, either an array-specific default or general color map
default, it would reset first to that site setting, and a second reset
would reset it to the XML defaults. Fixed this so that multiple
resetting of a color map always restores it to the appropriate site
settings color map.