Avoid initing MPI in Python interpreter.
When`paraview` is imported in a standard Python interpreter (not via
pvpython), we ended up initializing MPI due to !1987. That caused
failures generating sphinx documentation with Python 3. In any cause, it
sounds like a risky move to init MPI in a standard Python interpreter
unless user explicitly asked for. They do can that by doing the
following:

    import paraview
    paraview.options.batch = True