Cleanup plugin build and install locations.
ParaView plugins are now built and installed under the following
locations:

* Unix: lib/paraview-${ver}/plugins/
* OsX (app): <app>/Contents/Plugins/
* Windows: bin/plugins/

Plugins are also placed in a directory with the same name as the plugin.

`vtkPVPluginTracker` uses `vtkResourceFileLocator` to locate the plugins
at these new locations relative to both the ParaView libraries (first)
and current executable (second).

This fixes paraview/paraview#17832.