Fix screenshots with empty view.
When a multiview layout had an empty view, saving an multview image from
that failed. This was because as `vtkSMViewLayoutProxy::CaptureWindow`
was called, it would fire `ConfigureEvent` causing
`pqMultiViewWidget::reload` to update the pqQVTKWidget's to resize
messing up with the ViewSize that had been setup. This fixes it.

We don't need to call UpdateState while capturing image since the state
is not expected to change then anyways. This reverts ece2e86.
Updated SaveSmallScreenshot test to also test for the case ece2e86 was
fixing.

Fixes #17651.