Improved default extension for "save data".
Save data dialog will present "pvd" as the default extension if
available. For table datasets, it offers "csv" as the default.
Furthermore, when user saves the data of particular dataset type, we remember
the chosen extension. That way, next time the user tries
to save the data of same type, he/she will be offered the last chosen
extension.

Fixes #17538.