New QVTKOpenGLWidget Implementation with Cristal Eyes Stereo Support
A new implementation of the QVTKOpenGLWidget have been added
to VTK and this commit make uses of it in ParaView.

The main feature of this new widget is the support
of quad buffer based stereo mode "Cristal Eyes"
wich was unsupported since the switch to Qt5.

This new widget ensure the rendering can only
happen with a valid widget (aka ready for rendering)

This new widget support HiDpi rendering

This new widget improve also a few issues of
flickering with the old widget.

The new QVTKOpenGLWidget class does not support Qt::WA_NativeWindow flag
but the QScrollArea used in pqTransferFunctionWidget force it to be native.

The old widget have been kept and renammed
QVTKOpenGLSimpleWidget and should be used
for simple rendering or when the widget
is automatically native.
The old widget is used in pqTransferFunctionWidget

Depends on vtk/vtk!4317
Fix #16855
Fix #17740

Due to QTBUG-61836 (see QVTKOpenGLWidget::testingEvent()), events should
be propagated back to the internal QVTKOpenGLWindow when being fired
explicitely on the widget instance. We have to use a custom event
callback in this case to ensure that events are passed to the window.