Improve positioning of horizontal title when in vertical orientation
When the color legend is in vertical orientation and the title is
forced to be horizontal, change the anchoring points for left, center,
and right justifications to be relative to the bounding box of the
color bar AND the axis (which includes ticks and labels). This
improves the title position relative to the rest of the color legend.
Note that annotations are not considered when positioning the title.