Hide inapplicable filters from context menu
In the Filters menu in the menu bar, inapplicable filters are disabled
but still shown. In the pipeline browser context menu, we want them to
not be shown at all. This commit adds a flag to the
pqFiltersMenuReaction that determines whether disabled menu actions
should also be hidden. This flag is set to true for the context menu,
false for the menu bar.