Replace notify() call with sendEvent()
pqFlatTreeView would cause an infinite recursion for mouse wheel
events when the vertical scroll bar was visible. I suspect this was
happening because the wheelEvent() handler would call notify() to pass
the event to the scroll area. However, notify() propogates events up
to the parent of the receiver. Since the parent of the receiver was
the caller of notify() itself, the wheel event handler would be
invoked again, leading to the infinite recursion.

Using sendEvent() as was originally used solves the stack overflow
crash.