Hide Ghost Level option in GUI for writing unstructured grids and polydata
The Ghost Level option for writing out partitioned unstructured
grids and polydata is misleading so we hide that in the GUI so
that by default all available ghost information is written out.
If more ghost information is required then the Ghost Cells Generator
filter should be used and if not ghost information is wanted then
the Remove Ghost Information filter should be used. Note that the
Ghost Level option will still be available in Python.

Fixes #17294.