Changing output filename for gridwriter.py to input_%t.<extension>
In the upcoming PV changes we will try to hide the input name better
from the user so that they don't have to understand the concept of
multiple inputs. In that regard we're going to go with the convention
for the grid writer scripts to have the filename represent the
input name.