Avoid using obsolete data bounds.
Due to the disconnect between the representation and its mapper, when
`vtkPVRenderView::SynchronizeGeometryBounds` is called, there was a
potential that we'd end up using bounds from obsolete data. That was
causing #17941.

This fixes #17941 by ensuring that vtkPVDataDeliveryManager cleans any
obsolete datasets when a new one is provided to it by the
representation.