Handle null string correctly in vtkSIStringVectorProperty.
vtkSIStringVectorProperty attempted to handle null strings by using a
codeword. However vtkSMStringVectorProperty didn't handle it and the
codeword, "Invalid result", was being treated as the actual value for
the property.

There is no need for this. vtkSMStringVectorProperty doesn't distinguish
between null strings and empty strings and hence we don't need to have
this distinction in vtkSIStringVectorProperty as well.

Fixes #17946.