new catalyst/cinema export GUI.
The idea is a central panel for the user to declare what they want to
export, separated from the action of doing the export. The export
configuration is saveable and restorable in state files. Combined these
make it easier for the user to make Catalyst exports, avoid
some annoying behaviors, and open up paths to some new features.

The implementation uses savescreenshot and writerproxies instead of Catalyst
specific export proxies. We use standard proxies rather than Catalyst specific
ones to support new "export now" and "export batch script that will export"
use cases. The proxies are managed in something called an ExportProxyDepot for
more convenient access by the code.