Add support for multiple decorators.
Cleans up support for multiple decorators on the same property in
pqProxyWidget. Previously, it only worked if the decorators were of
different types. Now, muliple decorators of the same time are supported.
If multiple decorators are specified, an "and" operation is assumed for
both enabled state and visibility checks.

This fixes ui interactions referred to in #17438.