Modify vtkSMProxySelectionModel to uniquely add items
In commit ca54489a8c8 vtkSMProxySelectionModel was changed to use
an std::list instead of an std::set. This change didn't check for existing
copies of proxies before adding them to the list (which set did
automatically). Modified to make sure that multiple copies of the
same proxy are not added to the list.

Fixes #17961.