promote catalyst script generator features and unify in GUI
This moves the script generator dialog and the writers menu
from the plugin to inside the Catalyst menu. It will now always
be available, unless the GUI is built without python.

Changes to the SGWritersMenuManager essentially make it so that
I can use it to add the writers inside a menu rather than
owning an entire menu. In the temporalparallelism plugin it
still owns the entire menu.

to do: promote the tests, describe in the docs