Format-specific options for save-animation
Save animation dialog now shows format specific options. This allows
for the user to explicitly specify the format specific parameters like
quality, compression etc. To support this, pqSaveAnimationReaction now
prompts for the filename before showing the options dialog. Since that
conflicts with the "disconnect-and-save" option, we've removed that
option.

By cleaning up the defaults, this fixes #17952.