A hack way to preserve client-side animation/screenshot writers.
Typically, when we connect to a server, the client's proxy definitions
need to be replaced by those at the server, except, for client-only
proxies. While we have not taken care of this since the beginning, we
now have an issue for animation writers, for example, where if client
supports AVI and server doesn't, we still want the client able to save
AVIs. This hacks explicitly preserves animation and screenshot writers.