Changes Clip's InsideOut default from disabled to enabled
We change the default for Clip's InsideOut from disabled to enabled.
The reason for this is that for the box, cylinder and sphere clip
with Inside Out disabled it often looked like nothing happend because
the interior was clipped away. The InsideOut isn't about interior vs.
exterior but "keep" the section that that Normal points to (ie. in)
vs. keep the other section (out). For the box, sphere and cylinder
the normal points out. This commit also attempts to improve the
documentation.

Addresses #17438.