Make redistribution mode customizible.
We have come across representations that need different ways of handling
boundary cells when data is redistributed for ordered compositing.
Now a new API on vtkPVRenderView can be used by representation to mark
its data to be handle boundary cells differently. Default is still to
split the cells. A representation can choose to have those cells
duplicated or assigned to a unique region instead.