Making the --state command line option to be GUI only
The --state command line option could be passed to either paraview
or pvpython but it's only used in the GUI. This also updates the
--help output for both of them so that for pvpython it's no longer
given as an option. For the GUI it still is there but now also
notifies the user that a .py file can be passed in as well.