Change to links for ParaView Guide and Tutorial PDFs
Previously, these were shipped with the ParaView package, which meant
the PDFs needed to be updated prior to final release. Due to
scheduling constraints, it would not always be possible to update
these PDFs prior to release.

Instead, change the menu item for the guide and tutorial to be
external URLs pointing to a ParaView files directory on the web.  This
will make it possible to update the guides after the release and will
have the added benefit of making the installers smaller in size.

(cherry picked from commit 1401701379b059ddc4cc5011e7384861b88e5395)