File Dialog: issues with opening directories
File dialog was too eager to open directories rather than navigate to
them. That causes issues such as #18018. This change cleans up the UX.

When file dialog is in a mode that supports opening directories, we now
show a Navigate button that gets enabled if the selected item is a
directory. If Navigate is enabled/shown then it is also the default
button i.e. it will be the button that gets triggered when the user hits
the Enter key. This keeps the user-experience largely unchanged from
past where directories were concerned.

Fixes #18018.