Add XML_NAME_REGEX option to add_pqproxy macro
The add_pqproxy macro performs a string comparison on the xml group
and xml name of a proxy in order to create a user-defined pqProxy.
This commit adds an alternative to the xml name (XML_NAME_REGEX)
for comparison. If set, the generated code checks to see if the
proxy's name is captured by the xml name regular expression, rather
than performing a direct string comparison.