improve orientation in cinema spec A import
Addresses #18042

The major problem was that images in cinema store are flipped in Y.
At the same time I took out the rotation around Z to more closely
match the original roll. Rotation is nice, but when the
exported image has anything in the annotation layer (i.e. the colorbar)
the annotation is rotated too and the result is really distracting.

Note that there are still image clipping issues to address, to
demonstrate make the view much larger than the original and the
scale will right but the image will be clipped.

There is also the fundamental problem that the phi/theta camera
is fundamentally inadequate.

(cherry picked from commit 010f94c9f3fdb7caedcc6f755a0072969bef84bd)