Make array selection use pqTreeView
pqTreeView is better capable of support large number of items and user
interaction modes for toggling check states for multiple items. Hence,
this commit changes all array-status properties to use a
pqTreeView-based widget rather that the old pqTreeWidget-based widget.

For that, we add a new class pqArraySelectionWidget which replaces the
old (and poorly named) pqExodusIIVariableSelectionWidget.